refactor: update Group import path for consistency and remove unused addNavLinks method
